bug 10: switch to %~dp0, with fallback to the previous %0\.., and add a 2nd fallback of %PC2INSTALL%\bin
svn path=/trunk/; revision=1135

+++ bin/pc2reset.bat
@@ -6,6 +6,30 @@ rem
 rem $HeadURL$
 rem
 
+rem Windows 2000 and beyond syntax
+set PC2BIN=%~dp0
+if exist %PC2BIN%\pc2env.bat goto :continue
+
+rem fallback to path (or current directory)
+set PC2BIN=%0\..
+if exist %PC2BIN%\pc2env.bat goto :continue
+
+rem else rely on PC2INSTALL variable
+set PC2BIN=%PC2INSTALL%\bin
+if exist %PC2BIN%\pc2env.bat goto :continue
+
+echo.
+echo ERROR: Could not locate scripts.
+echo.
+echo Please set the varible PC2INSTALL to the location of
+echo   the VERSION file (ex: c:\pc2-9.0.0)
+echo.
+pause
+goto :end
+
+:continue
+call %PC2BIN%\pc2env.bat
+
 set RMCMD=rmdir /s /q
 if %OS%. == . set RMCMD=deltree /y
 
@@ -13,7 +37,7 @@ if not exist logs goto wrongdir
 if not exist packets goto wrongdir
 
 echo Backing up settings to archive 
-call %0\..\pc2zip.bat
+call %PC2BIN%\pc2zip.bat
 
 %RMCMD% logs
 %RMCMD% packets
